try MVC to divide code and hide key in .env
